Add unknown property error handling
There was a problem when polling a point which had no
description.

In that case, bacpypes returns an unknownProperty
and BAC0 was skipping the value. Skipping the value
would lead to BAC0 taking the name of the next point
as the description of the current one when putting
points in batches. From this point, all following
points would be wrong.